migrations: change assessment_value data type

+++ internal/storage/postgres/migrations/20240427093228_init_table.up.sql
@@ -46,7 +46,7 @@ CREATE TABLE assessment(
     id UUID not null,
     users_id UUID not null,
     courses_id UUID not null,
-    assessment_value int,
+    assessment_value FLOAT,
     assessment_code varchar(10),
     created_at TIMESTAMP,
     updated_at TIMESTAMP,
